refactor: unify db layer
